backend/requirements.txt

langchain==0.2.1
chromadb==0.5.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain-google-genai==1.0.6
pypdf==4.2.0
langchain-community==0.2.1
